+++ .gitignore
@@ -1,1 +1,2 @@
 *.log
+node_modules/
