+++ package.json
@@ -3,6 +3,9 @@
   "version": "2.0.5",
   "description": "Knayi Myanmar languages JavaScirpt library",
   "main": "main.js",
+  "scripts":{
+  	"test":"mocha"
+  },
   "repository": {
     "type": "git",
     "url": "http://github.com/greenlikeorange/knayi-myscript"
@@ -23,5 +26,9 @@
   "bugs": {
     "url": "https://github.com/greenlikeorange/knayi-myscript/issues"
   },
-  "homepage": "https://github.com/greenlikeorange/knayi-myscript"
+  "homepage": "https://github.com/greenlikeorange/knayi-myscript",
+  "devDependencies": {
+    "chai": "^4.0.2",
+    "mocha": "^3.4.2"
+  }
 }
